Rework entire repo architecture
