+++ src/evaluation/xai_utils.py
@@ -12,12 +12,12 @@ from torch import optim
 import torch.nn as nn
 from torch.nn.utils.rnn import pack_padded_sequence, pad_packed_sequence
 
-from preprocess_utils import *
-from train import train_model, test_model
-from models import BasicLSTM, BiLSTM
-from test_save_stats import *
+from src.utils.preprocess_utils import *
+from src.training.train_utils import train_model, test_model
+from src.models import BasicLSTM, BiLSTM
+from src.evaluation.test_save_stats import *
 
-from utils import *
+from src.utils.utils import *
 
 import captum
 from captum.attr import LayerIntegratedGradients, TokenReferenceBase, visualization
